Restructure networking code; fix errors caused by CAN overhaul
