+++ src/Globals.cpp
@@ -9,7 +9,7 @@
 namespace Globals {
 CommandLineOptions opts;
 RoverState curr_state;
-websocket::SingleClientWSServer websocketServer("DefaultServer", Constants::WS_SERVER_PORT);
+net::websocket::SingleClientWSServer websocketServer("DefaultServer", Constants::WS_SERVER_PORT);
 std::atomic<bool> E_STOP = false;
 std::atomic<bool> AUTONOMOUS = false;
 } // namespace Globals
