--- src/Networking/NetworkingStubs.cpp
@@ -1,57 +1,0 @@
-
-#include "NetworkingStubs.h"
-
-#include "../Globals.h"
-
-#include <queue>
-
-std::queue<CANPacket> can_packets{};
-std::queue<std::string> base_station_packets{};
-
-int numCANPackets() {
-	return can_packets.size();
-}
-
-int numBaseStationPackets() {
-	return base_station_packets.size();
-}
-
-void clearTestGlobals() {
-	while (!can_packets.empty()) {
-		can_packets.pop();
-	}
-	while (!base_station_packets.empty()) {
-		base_station_packets.pop();
-	}
-	Globals::status_data = nlohmann::json();
-}
-
-//// Commented out to avoid a duplicate definition. I know it's not going to be
-////   correct but all these tests are going to get superseded soon anyway
-//   void sendCANPacket(const CANPacket& packet) {
-//   	can_packets.push(packet);
-//   }
-
-void sendBaseStationPacket(const std::string& packet) {
-	base_station_packets.push(packet);
-}
-
-int recvBaseStationPacket(char* buffer) {
-	return 0;
-}
-
-bool InitializeBaseStationSocket() {
-	return true;
-}
-
-CANPacket popCANPacket() {
-	CANPacket p = can_packets.front();
-	can_packets.pop();
-	return p;
-}
-
-std::string popBaseStationPacket() {
-	std::string p = base_station_packets.front();
-	base_station_packets.pop();
-	return p;
-}
