--- src/Networking/NetworkingStubs.h
@@ -1,17 +1,0 @@
-
-#pragma once
-
-extern "C" {
-#include <HindsightCAN/CANPacket.h>
-}
-#include <string>
-
-// For testing purposes. The implementations of sendCANPacket and sendBaseStationPacket in
-// NetworkingStubs.cpp will just save their arguments into an array
-// which can be retrieved using popCANPacket and popBaseStationPacket
-
-int numCANPackets();
-int numBaseStationPackets();
-void clearTestGlobals();
-CANPacket popCANPacket();
-std::string popBaseStationPacket();
