--- src/Networking/ParseBaseStation.h
@@ -1,10 +1,0 @@
-
-#ifndef __PARSE_BASE_STATION_H__
-#define __PARSE_BASE_STATION_H__
-
-#include <nlohmann/json.hpp>
-using nlohmann::json;
-bool ParseBaseStationPacket(char const* buffer);
-bool sendError(std::string const& msg);
-
-#endif
