+++ src/network/websocket/WebSocketProtocol.cpp
@@ -1,8 +1,8 @@
 #include "WebSocketProtocol.h"
 
-#include "../../log.h"
 #include "../../Constants.h"
-
+#include "../../log.h"
+namespace net {
 namespace websocket {
 
 static const std::string TYPE_KEY = "type";
@@ -64,7 +64,7 @@ void WebSocketProtocol::processMessage(const json& obj) const {
 		auto validatorEntry = validatorMap.find(messageType);
 		if (validatorEntry != validatorMap.end()) {
 			if (validatorEntry->second(obj)) {
-				if(protocolPath == Constants::MC_PROTOCOL_NAME){
+				if (protocolPath == Constants::MC_PROTOCOL_NAME) {
 					log(LOG_INFO, "MC->R: %s\n", obj.dump().c_str());
 				}
 				handlerMap.at(messageType)(obj);
@@ -87,3 +87,4 @@ std::string WebSocketProtocol::getProtocolPath() const {
 }
 
 } // namespace websocket
+} // namespace net
